+++ BLOCHUS/BLOCHUS.m
@@ -33,10 +33,10 @@ function BLOCHUS
 %------------- BEGIN CODE --------------
 
 %% GUI 'header' info and default GUI settings
-myui.version = '0.1.4';
-myui.date = '25.09.2020';
+myui.version = '0.1.5';
+myui.date = '30.06.2022';
 myui.author = 'Thomas Hiller';
-myui.email = 'thomas.hiller[at]leibniz-liag.de';
+myui.email = 'thomas.hiller[at]bgr.de';
 
 myui.fontsize = 9;
 myui.axfontsize = 11;